fix discrepancy of return type of 'strlen'

+++ pygraphviz/graphviz_wrap.c
@@ -3022,6 +3022,7 @@ static swig_module_info swig_module = {swig_types, 11, 0, 0, 0, 0};
 
 #include "graphviz/cgraph.h"
 #include "graphviz/gvc.h"
+#include <stddef.h>
 
 
 SWIGINTERN swig_type_info*
@@ -3405,7 +3406,7 @@ SWIG_FromCharPtr(const char *cptr)
    * @return The equivalent of `val`, as a plain string or HTML-like string, as relevant
    */
   static char *htmlize(Agraph_t *g, const char *name, char *val) {
-    int len;
+    size_t len;
     char *hs;
 
     if (val[0] == '<' && (strcmp(name, "label") == 0 || strcmp(name, "xlabel") == 0)) {
